// src/theme/types.ts
export type Breakpoints = string[];

export type MediaQueries = {
  fold: string;
  xxs: string;
  xs: string;
  sm: string;
  md: string;
  tab: string;
  lg: string;
  xl: string;
  nav: string;
};

export type Spacing = number[];

export type Radii = {
  small: string;
  default: string;
  card: string;
  circle: string;
};

export type Shadows = {
  popper: string;
  popperSecondary: string;
};

export type Gradients = {
  bubblegum: string;
  cardHeader: string;
  blue: string;
  violet: string;
  violetAlt: string;
};

export type Colors = {
  bodyBackground: string;
  blockBG: string;
  text: string;
  textAccent: string;
  purple: string;
  red: string;
  green: string;
  darkGreen: string;
  statusGreen: string;
  statusPink: string;
  gold: string;
  
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  tertiary: string;
  success: string;
  failure: string;
  warning: string;
  background: string;
  darkGray:string;
  backgroundInvert: string;
  backgroundDisabled: string;
  backgroundAlt: string;
  particlesBckground: string;
  white: string;
  grey: string;
  drawerBackground: string;
  gradientBackground: string;
  gradientBackgroundSecondary: string;
  protocolsMarqueeBackground: string;
  scrollbarTrackBorder: string;
  scrollbarThumbBg: string;
  drawerSubMenuBackgroud: string;
  adminDashboadText: string;
  apple: string;
  pastelGreen: string;
  modalText: string;
  colorInvert: string;
};

export type ZIndices = {
  dropdown: number;
  modal: number;
};

export type Fonts = {
  xs: number;
  sm: number;
  md: number;
  lg: number;
  xl: number;
};

export type Layout = {
  background: string;
  padding: string;
  margin: string;
};